pyproject.toml
[project]
name = "mb-db"
version = "0.1.0"
description = "Millenium blades wiki"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "joblib>=1.5.1",
    "keras-ocr>=0.9.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "tensorflow>=2.19.0",
]
